backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
redis
httpx
python-jose[cryptography]
python-dotenv
